Allocation Plan API
- domain refactor
- added REST API controller method
- refactored DTS model for basic validation
- improved mapper functions for ease of use
- code cleaning

+++ src/main/go/domain/allocation_plan.go
@@ -35,7 +35,7 @@ type AllocationPlan struct {
 	AllocationPlanIdentifier
 	PlanType             allocation.PlanType
 	PlannedExecutionDate *time.Time
-	PortfolioId          int
+	PortfolioId          int64
 	Details              []*PlannedAllocation
 }
